chore(deals): the base currency gets its own file, and the migration its own number
0189 was already taken on main by the site-read logo migration, and 0190 by
the anchor lifecycle guard, so the canonical-company-values migration moves
to 0191.

fxrate_store.go had grown to 519 lines and failed the 500-line gate. The base
currency splits out whole: it is what the sheet converts INTO, not a rate in
it.
